src/main/java/com/example/gestionempleados/controller/EmpleadoController.java
package com.example.gestionempleados.controller;

import com.example.gestionempleados.entity.Empleado;
import com.example.gestionempleados.service.EmpleadoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/empleados")
public class EmpleadoController {

    @Autowired
    private EmpleadoService empleadoService;

    @PostMapping
    public Empleado addEmpleado(@RequestBody Empleado empleado) {
        return empleadoService.saveEmpleado(empleado);
    }

    @GetMapping
    public List<Empleado> getAllEmpleados() {
        return empleadoService.getAllEmpleados();
    }

    @GetMapping("/salario-total")
    public Double getSalarioTotal() {
        return empleadoService.calcularSalarioTotal();
    }
}